Make init-in-transaction? fix less intrusive
The earlier fix introduced in the commit dfe79461 was problematic in the sense
that it injected parameters into the `:db` configuration field assuming that
the value was a map.  In addition to db-spec maps this field is supposed to
accept also objects that implement the `Connection` or `DataSource` interface.
Thus the original change breaks the library for any configuration using these.

This change attempts to revert the intrusive parts of the original change
while still addressing the issue of `:init-in-transaction?` not being honored
when set to false.